+++ fdbserver/DDTeamCollection.actor.cpp
@@ -1060,8 +1060,7 @@ public:
 		state Future<Void> failureTracker;
 		state ServerStatus status(server->getLastKnownInterface().locality);
 		state bool lastIsUnhealthy = false;
-		state Reference<IDDTxnProcessor> txnProcessor = makeReference<DDTxnProcessor>(cx);
-		state Future<Void> metricsTracker = server->serverMetricsPolling(txnProcessor);
+		state Future<Void> metricsTracker = server->serverMetricsPolling(self->db);
 
 		state Future<std::pair<StorageServerInterface, ProcessClass>> interfaceChanged = server->onInterfaceChanged;
 		state bool hasWrongDC = !self->isCorrectDC(*server);
