+++ fdbserver/include/fdbserver/DataDistribution.actor.h
@@ -169,33 +169,6 @@ private:
 	    moveReason(DataMovementReason::INVALID) {}
 };
 
-struct GetTeamRequest;
-namespace data_distribution {
-class EligibilityCounter {
-public:
-	enum Type {
-		NONE = 0, // don't care about eligibility
-		LOW_CPU = 1,
-		LOW_DISK_UTIL = 2
-	};
-
-	// set the count of type to 0
-	void reset(Type type);
-
-	// return the minimal count of a combined eligible type
-	unsigned getCount(int combinedType) const;
-
-	// increase the count of type
-	void increase(Type type);
-
-	static int fromGetTeamRequest(GetTeamRequest const&);
-
-private:
-	std::map<Type, unsigned int> type_count;
-};
-
-} // namespace data_distribution
-
 struct GetMetricsRequest {
 	KeyRange keys;
 	Promise<StorageMetrics> reply;
